Fix vendor go dependency management
