+++ .gitignore
@@ -1,5 +1,4 @@
 .env
-vendor
 .idea
 elm-stuff
 .idea/
