+++ backend/goPaths.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+export GOPATH=$PWD
+export PATH=$PATH:$GOPATH/bin
