--- elm-client/Model.elm
@@ -1,5 +1,0 @@
-module Model exposing (Task)
-
-type alias Task = {
-    description : String
-}
